% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPos.R
\name{createPos}
\alias{createPos}
\title{Create a 2-column matrix from two vectors}
\usage{
createPos(X, Y)
}
\arguments{
\item{X}{one array of length n}

\item{Y}{another array of length n}
}
\value{
A 2-column numerical matrix
}
\description{
Create a 2-column matrix from two vectors
}
\examples{
createPos(c(1,2,3), c(1,4,9))
}
