% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorClass.R
\docType{class}
\name{Sensor-class}
\alias{Sensor-class}
\title{An S4 class to represent a sensor}
\description{
An S4 class to represent a sensor
}
\section{Slots}{

\describe{
\item{\code{ox}}{A 2xN numeric matrix, 
representing [x, y] of sensor in oxidized state}

\item{\code{red}}{A 2xN numeric matrix, 
representing [x, y] of sensor in reduced state}

\item{\code{E0}}{a numeric value for the midpoint potential of the sensor}
}}

