% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorClass.R
\name{Sensor}
\alias{Sensor}
\title{A function to construct a Sensor class based on two matricies}
\usage{
Sensor(oxMatrix, redMatrix, E0)
}
\arguments{
\item{oxMatrix}{A 2XN numeric matrix}

\item{redMatrix}{A 2xN numeric matrix}

\item{E0}{a numeric value for the midpoint potential of the sensor}
}
\value{
A sensor object
}
\description{
A function to construct a Sensor class based on two matricies
}
