% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensorClass.R
\name{SensorValues}
\alias{SensorValues}
\title{A function to construct a sensor class based on sensor properties}
\usage{
SensorValues(Rmax, Rmin, delta,
lambda_1 = 410, lambda_2 = 470, E0 = -265)
}
\arguments{
\item{Rmax}{the maximum ratiometric intensity value}

\item{Rmin}{the minimum ratiometric intensity value}

\item{delta}{the dynamic range of the divisor wavelength}

\item{lambda_1}{(optional) dividend wavelength, default = 410}

\item{lambda_2}{(optional) divisor wavelength, default = 470}

\item{E0}{(optional) midpoint potential of the sensor, 
default=-265}
}
\value{
A sensor object
}
\description{
A function to construct a sensor class based on sensor properties
}
